+++ src/MEL/Communications/TcpSocket.cpp
@@ -206,7 +206,7 @@ Socket::Status TcpSocket::send(const void* data, std::size_t size, std::size_t& 
     for (sent = 0; sent < size; sent += result)
     {
         // Send a chunk of data
-        result = ::send(get_handle(), static_cast<const char*>(data) + sent, size - sent, flags);
+        result = ::send(get_handle(), static_cast<const char*>(data) + sent, static_cast<int>(size - sent), flags);
 
         // Check for errors
         if (result < 0)
